add my course feature
