+++ lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_alterra_app/feature/home/course_list_page/view/ui/flutter_course_list_page.dart';
-import 'package:flutter_alterra_app/feature/home/course_list_page/view/ui/ui_ux_course_list_page.dart';
 
+import 'feature/home/my_course_page/view/ui/my_course_page.dart';
 
 import 'feature/auth/login/view/ui/splash_page.dart';
 
@@ -19,7 +18,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         fontFamily: 'Inter',
       ),
-      home: const UiUxCourseListPage(),
+      home: const MyCoursePage(),
     );
   }
 }
